ssr & reformat
