+++ layouts/simplified.vue
@@ -3,11 +3,7 @@ import { nameI18n } from '~/utils/i18n';
 
 const { locale } = useI18n()
 const localePath = useLocalePath()
-const basicInfo = useDataMerchantBasic()
-
-onMounted(() => {
-  stateMerchant.basic();
-});
+const basicInfo = await stateMerchant.basic(true)
 </script>
 
 <template>
