+++ server/middleware/entrance.js
@@ -44,7 +44,7 @@ const wrapMapi = async (event) => {
 
     // custom mapi cache rule
     if (['/api/merchant/basicInfo', '/api/merchant/info'].includes(purePath)) {
-      event.idatariverReq.cacheKey = purePath // without query params
+      event.idatariverReq.cacheKey = `${purePath}-${getHeader(event, 'X-Idr-Locale')}` // without query params
     } else if (['/api/order/search'].includes(purePath)) {
       event.idatariverReq.cacheKey = event.path // with query params
     }
